requirements/base.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.1
django-environ==0.8.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.4.2
identify==2.4.7
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
nodeenv==1.6.0
openapi-codec==1.3.2
packaging==21.3
platformdirs==2.4.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
